Rename permutational package to stepdependencies, extract PERMUTATIONAL from core (#237)
Groundwork for issue #224 (multi-module split), continuing from the merged
MutationMode SPI decoupling PR.

- Rename com.adobe.campaign.tests.integro.phased.permutational to
  ...stepdependencies. The package (ScenarioStepDependencies,
  ScenarioStepDependencyFactory, StepDependencies) was built in 2022 purely
  for legacy Phased Testing's step-order detection; "permutational" only
  describes half its purpose and its accidental echo of "Mutational" made
  it look mutational-owned, when it isn't.

- Extract the ExecutionMode.PERMUTATIONAL execution mode out of
  PhasedTestManager: it was a by-product of the stepdependencies research,
  but is exclusively consumed by Mutational today, and was hardcoded
  inline in a core method (the same anti-pattern the merged SPI PR
  eliminated elsewhere, missed in that sweep since it didn't reference
  MutationManager by name). PhasedTestManager.fetchProvidersShuffled is
  restored to its original INTERRUPTIVE/NON_INTERRUPTIVE-only scope;
  MutationManager gains its own fetchProvidersShuffled that handles
  PERMUTATIONAL and delegates everything else back to PhasedTestManager.

- Move the two PhasedTestManagerTests that exercised this PERMUTATIONAL
  behavior (and were already reaching into mutational-only test fixtures
  to do it) into MutationManagerTests, alongside the code they test.
