+++ src/main/java/com/adobe/campaign/tests/integro/phased/PhasedTestListener.java
@@ -11,9 +11,9 @@ package com.adobe.campaign.tests.integro.phased;
 import com.adobe.campaign.tests.integro.phased.exceptions.PhasedTestConfigurationException;
 import com.adobe.campaign.tests.integro.phased.exceptions.PhasedTestException;
 import com.adobe.campaign.tests.integro.phased.internal.PhaseProcessorFactory;
-import com.adobe.campaign.tests.integro.phased.permutational.ScenarioStepDependencies;
-import com.adobe.campaign.tests.integro.phased.permutational.ScenarioStepDependencyFactory;
-import com.adobe.campaign.tests.integro.phased.permutational.StepDependencies;
+import com.adobe.campaign.tests.integro.phased.stepdependencies.ScenarioStepDependencies;
+import com.adobe.campaign.tests.integro.phased.stepdependencies.ScenarioStepDependencyFactory;
+import com.adobe.campaign.tests.integro.phased.stepdependencies.StepDependencies;
 import com.adobe.campaign.tests.integro.phased.utils.ClassPathParser;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
